requirements.txt
qdrant-client>=1.11.0
pyyaml>=6.0.1
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0
python-dotenv>=1.0.1
